requirements.txt
accelerate==0.33.0
annotated-types==0.7.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorlog==6.8.2
deepspeed==0.14.4
filelock==3.13.1
fsspec==2024.2.0
hjson==3.1.0
huggingface-hub==0.24.5
idna==3.4
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.2
numpy==1.26.4
nvidia-ml-py==12.555.43
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow==10.2.0
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.28.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchvision==0.16.1+cu121
tqdm==4.66.5
transformers==4.44.0
triton==2.1.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
